# packages/shared-contracts/openapi/cloudhelm.openapi.yaml
openapi: 3.1.0
info:
  title: CloudHelm Platform API
  description: CloudHelm 平台 API。M2 提供真实数据库驱动的项目、任务、需求、设计、审批和事件底座。
  version: 0.2.0
paths:
...
  /api/tasks:
    post:
      tags:
      - tasks
      summary: 创建任务
      description: 创建 Task，校验项目存在，并写入 TaskCreated 事件。
      operationId: create_task_api_tasks_post
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskCreate'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskRead'
        '400':
          description: 业务请求错误。
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: 资源不存在。
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: 状态冲突。
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: 请求参数校验失败。
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: 服务端错误。
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    get:
      tags:
      - tasks
      summary: 分页读取任务
      description: 分页读取任务列表。
      operationId: list_tasks_api_tasks_get
      parameters:
      - name: project_id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          description: 按项目过滤。
          title: Project Id
        description: 按项目过滤。
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 100
          minimum: 1
          description: 分页大小。
          default: 50
          title: Limit
        description: 分页大小。
      - name: cursor
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: offset cursor。
          title: Cursor
        description: offset cursor。
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PageResponse_TaskRead_'
        '400':
          description: 业务请求错误。
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: 资源不存在。
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: 状态冲突。
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: 请求参数校验失败。
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: 服务端错误。
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/tasks/{task_id}:
    get:
      tags:
      - tasks
      summary: 读取任务详情
      description: 读取单个 Task。
      operationId: get_task_api_tasks__task_id__get
      parameters:
      - name: task_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Task Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskRead'
        '400':
          description: 业务请求错误。
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: 资源不存在。
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: 状态冲突。
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: 请求参数校验失败。
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: 服务端错误。
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/tasks/{task_id}/pause:
    post:
      tags:
      - tasks
      summary: 暂停任务
      description: 暂停任务并写入 TaskPaused 事件。
      operationId: pause_task_api_tasks__task_id__pause_post
      parameters:
      - name: task_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Task Id
      requestBody:
        content:
          application/json:
            schema:
              anyOf:
              - $ref: '#/components/schemas/DecisionRequest'
              - type: 'null'
              title: Payload
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskRead'
        '400':
          description: 业务请求错误。
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: 资源不存在。
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: 状态冲突。
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: 请求参数校验失败。
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: 服务端错误。
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/tasks/{task_id}/resume:
    post:
      tags:
      - tasks
      summary: 恢复任务
      description: 恢复暂停任务并写入 TaskResumed 事件。
      operationId: resume_task_api_tasks__task_id__resume_post
      parameters:
      - name: task_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Task Id
      requestBody:
        content:
          application/json:
            schema:
              anyOf:
              - $ref: '#/components/schemas/DecisionRequest'
              - type: 'null'
              title: Payload
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskRead'
        '400':
          description: 业务请求错误。
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: 资源不存在。
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: 状态冲突。
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: 请求参数校验失败。
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: 服务端错误。
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/tasks/{task_id}/cancel:
    post:
      tags:
      - tasks
      summary: 取消任务
      description: 取消任务并写入 TaskCancelled 事件。
      operationId: cancel_task_api_tasks__task_id__cancel_post
      parameters:
      - name: task_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Task Id
      requestBody:
        content:
          application/json:
            schema:
              anyOf:
              - $ref: '#/components/schemas/DecisionRequest'
              - type: 'null'
              title: Payload
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskRead'
        '400':
          description: 业务请求错误。
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: 资源不存在。
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: 状态冲突。
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: 请求参数校验失败。
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: 服务端错误。
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/tasks/{task_id}/requirements:
    post:
      tags:
      - requirements
      summary: 创建需求规格
      description: 保存真实需求规格，不自动生成或伪造 Agent 输出。
      operationId: create_requirement_api_tasks__task_id__requirements_post
      parameters:
      - name: task_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Task Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RequirementSpecCreate'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RequirementSpecRead'
        '400':
          description: 业务请求错误。
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: 资源不存在。
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: 状态冲突。
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: 请求参数校验失败。
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: 服务端错误。
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    get:
      tags:
      - requirements
      summary: 分页读取任务需求规格
      description: 读取某任务下的需求规格列表。
      operationId: list_requirements_api_tasks__task_id__requirements_get
      parameters:
      - name: task_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Task Id
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 100
          minimum: 1
          description: 分页大小。
          default: 50
          title: Limit
        description: 分页大小。
      - name: cursor
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: offset cursor。
          title: Cursor
        description: offset cursor。
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PageResponse_RequirementSpecRead_'
        '400':
          description: 业务请求错误。
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: 资源不存在。
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: 状态冲突。
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: 请求参数校验失败。
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: 服务端错误。
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/requirements/{requirement_id}:
    get:
      tags:
      - requirements
      summary: 读取需求规格
      description: 读取单个需求规格。
      operationId: get_requirement_api_requirements__requirement_id__get
      parameters:
      - name: requirement_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Requirement Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RequirementSpecRead'
        '400':
          description: 业务请求错误。
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: 资源不存在。
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: 状态冲突。
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: 请求参数校验失败。
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: 服务端错误。
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/requirements/{requirement_id}/approve:
    post:
      tags:
      - requirements
      summary: 通过需求规格
      description: 通过需求规格并写入 RequirementSpecApproved 事件。
      operationId: approve_requirement_api_requirements__requirement_id__approve_post
      parameters:
      - name: requirement_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Requirement Id
      requestBody:
        content:
          application/json:
            schema:
              anyOf:
              - $ref: '#/components/schemas/DecisionRequest'
              - type: 'null'
              title: Payload
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RequirementSpecRead'
        '400':
          description: 业务请求错误。
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: 资源不存在。
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: 状态冲突。
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: 请求参数校验失败。
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: 服务端错误。
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/requirements/{requirement_id}/request-changes:
    post:
      tags:
      - requirements
      summary: 要求修改需求规格
      description: 要求修改需求规格并写入事件。
      operationId: request_requirement_changes_api_requirements__requirement_id__request_changes_post
      parameters:
      - name: requirement_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Requirement Id
      requestBody:
        content:
          application/json:
            schema:
              anyOf:
              - $ref: '#/components/schemas/DecisionRequest'
              - type: 'null'
              title: Payload
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RequirementSpecRead'
        '400':
          description: 业务请求错误。
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: 资源不存在。
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: 状态冲突。
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: 请求参数校验失败。
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: 服务端错误。
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  schemas:
...
    DecisionRequest:
      properties:
        actor_id:
          type: string
          minLength: 1
          title: Actor Id
          description: 操作人或组件标识。
          default: user
        reason:
          anyOf:
          - type: string
          - type: 'null'
          title: Reason
          description: 操作原因。
      type: object
      title: DecisionRequest
      description: 审批、通过或要求修改类动作的请求体。
    ErrorResponse:
      properties:
        code:
          type: string
          title: Code
          description: 稳定错误码。
        message:
          type: string
          title: Message
          description: 面向调用方的错误描述。
        detail:
          anyOf:
          - {}
          - type: 'null'
          title: Detail
          description: 调试细节或字段错误。
        trace_id:
          type: string
          title: Trace Id
          description: 链路追踪 ID。
      type: object
      required:
      - code
      - message
      - trace_id
      title: ErrorResponse
      description: 统一错误响应结构。
...
    PageInfo:
      properties:
        limit:
          type: integer
          maximum: 100.0
          minimum: 1.0
          title: Limit
          description: 本次最多返回条数。
        next_cursor:
          anyOf:
          - type: string
          - type: 'null'
          title: Next Cursor
          description: 下一页游标；为空表示无更多数据。
      type: object
      required:
      - limit
      title: PageInfo
      description: '分页元数据。


        M2 使用 offset cursor 的最小实现，后续如需高并发滚动查询可升级为

        基于 `created_at,id` 的稳定游标。'
...
    PageResponse_RequirementSpecRead_:
      properties:
        items:
          items:
            $ref: '#/components/schemas/RequirementSpecRead'
          type: array
          title: Items
          description: 当前页数据。
        page:
          $ref: '#/components/schemas/PageInfo'
          description: 分页信息。
      type: object
      required:
      - items
      - page
      title: PageResponse[RequirementSpecRead]
    PageResponse_TaskRead_:
      properties:
        items:
          items:
            $ref: '#/components/schemas/TaskRead'
          type: array
          title: Items
          description: 当前页数据。
        page:
          $ref: '#/components/schemas/PageInfo'
          description: 分页信息。
      type: object
      required:
      - items
      - page
      title: PageResponse[TaskRead]
...
    RequirementSpecCreate:
      properties:
        source_type:
          type: string
          minLength: 1
          title: Source Type
          description: 需求来源类型。
          default: manual
        raw_input:
          type: string
          minLength: 1
          title: Raw Input
          description: 原始需求输入。
        user_story:
          anyOf:
          - type: string
          - type: 'null'
          title: User Story
          description: 用户故事。
        constraints_json:
          items: {}
          type: array
          title: Constraints Json
          description: 约束条件 JSON 数组。
        acceptance_criteria_json:
          items: {}
          type: array
          title: Acceptance Criteria Json
          description: 验收标准 JSON 数组。
      type: object
      required:
      - raw_input
      title: RequirementSpecCreate
      description: 创建需求规格的请求体。
    RequirementSpecRead:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        task_id:
          type: string
          format: uuid
          title: Task Id
        project_id:
          type: string
          format: uuid
          title: Project Id
        source_type:
          type: string
          title: Source Type
        raw_input:
          type: string
          title: Raw Input
        user_story:
          anyOf:
          - type: string
          - type: 'null'
          title: User Story
        constraints_json:
          items: {}
          type: array
          title: Constraints Json
        acceptance_criteria_json:
          items: {}
          type: array
          title: Acceptance Criteria Json
        status:
          $ref: '#/components/schemas/ReviewStatus'
        version:
          type: integer
          title: Version
        created_at:
          type: string
          format: date-time
          title: Created At
        updated_at:
          type: string
          format: date-time
          title: Updated At
      type: object
      required:
      - id
      - task_id
      - project_id
      - source_type
      - raw_input
      - user_story
      - constraints_json
      - acceptance_criteria_json
      - status
      - version
      - created_at
      - updated_at
      title: RequirementSpecRead
      description: 需求规格响应结构。
    ReviewStatus:
      type: string
      enum:
      - draft
      - approved
      - changes_requested
      title: ReviewStatus
      description: 需求和技术设计的人工评审状态。
    RiskLevel:
      type: string
      enum:
      - L0
      - L1
      - L2
      - L3
      - L4
      title: RiskLevel
      description: CloudHelm 工具和任务风险等级。
    TaskCreate:
      properties:
        project_id:
          type: string
          format: uuid
          title: Project Id
          description: 所属项目 ID，必须已存在。
        title:
          type: string
          maxLength: 160
          minLength: 1
          title: Title
          description: 任务标题。
        description:
          type: string
          minLength: 1
          title: Description
          description: 任务描述。
        source_type:
          type: string
          minLength: 1
          title: Source Type
          description: 任务来源类型。
          default: manual
        source_ref:
          anyOf:
          - type: string
          - type: 'null'
          title: Source Ref
          description: 来源引用。
        risk_level:
          $ref: '#/components/schemas/RiskLevel'
          description: 任务初始风险等级。
          default: L0
        created_by:
          type: string
          minLength: 1
          title: Created By
          description: 创建人或组件。
          default: user
      type: object
      required:
      - project_id
      - title
      - description
      title: TaskCreate
      description: 创建 Task 的请求体。
    TaskRead:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        project_id:
          type: string
          format: uuid
          title: Project Id
        title:
          type: string
          title: Title
        description:
          type: string
          title: Description
        source_type:
          type: string
          title: Source Type
        source_ref:
          anyOf:
          - type: string
          - type: 'null'
          title: Source Ref
        status:
          $ref: '#/components/schemas/TaskStatus'
        risk_level:
          $ref: '#/components/schemas/RiskLevel'
        current_phase:
          type: string
          title: Current Phase
        created_by:
          type: string
          title: Created By
        created_at:
          type: string
          format: date-time
          title: Created At
        updated_at:
          type: string
          format: date-time
          title: Updated At
      type: object
      required:
      - id
      - project_id
      - title
      - description
      - source_type
      - source_ref
      - status
      - risk_level
      - current_phase
      - created_by
      - created_at
      - updated_at
      title: TaskRead
      description: Task 响应结构。
    TaskStatus:
      type: string
      enum:
      - created
      - running
      - waiting_approval
      - paused
      - failed
      - done
      - cancelled
      title: TaskStatus
      description: 任务状态枚举，与设计文档数据层保持一致。
...
servers:
- url: http://127.0.0.1:18080
  description: 本地开发平台 API